cron/check-apt-key-expired.sh BETA
Problem: unattended upgrades do not upgrade gitlab, because apt-get
shows a warning and ignores the package repo.

  The following signatures were invalid: EXPKEYSIG ...

It should exit with an error instead, but it does not.

This script `/usr/local/sbin/check-apt-key-expired.sh` checks for
expired keys using `apt-key`.

Check `example.crontab` on how to use this in cron.

- https://unix.stackexchange.com/q/153037/120440
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=745735 debian issue since
  2014